--- lokalilmu_guru/lib/blocs/dashboard/dashboard_event.dart
@@ -1,9 +1,0 @@
-abstract class DashboardEvent {}
-
-class LoadDashboardEvent extends DashboardEvent {}
-
-class ToggleCourseStateEvent extends DashboardEvent {
-  final bool hasJoinedCourses;
-
-  ToggleCourseStateEvent(this.hasJoinedCourses);
-}
