--- lokalilmu_guru/lib/widgets/dashboard/empty_state_widget.dart
@@ -1,25 +1,0 @@
-import 'package:flutter/material.dart';
-
-class EmptyStateWidget extends StatelessWidget {
-  final String message;
-
-  const EmptyStateWidget({
-    Key? key,
-    required this.message,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: double.infinity,
-      padding: const EdgeInsets.symmetric(vertical: 16),
-      child: Text(
-        message,
-        style: const TextStyle(
-          color: Colors.grey,
-          fontSize: 14,
-        ),
-      ),
-    );
-  }
-}
